backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
openai
faiss-cpu
youtube-transcript-api
pydantic
tiktoken
